Implement abstract factory pattern
Abstract Factory Pattern has been implemented in this commit. It
is demonstrated on a list of devices that are further divided into
two categories - 'Laptop' and 'Phone'. The final result is that
if subcategory is known along with the device, then a runtime
instatiation of the devices can be performed.